// sed-frontend/src/app/shared/model/common/garante.ts
export class Garante {
    catalogo: string;
    codigo: string;
    codigopadre: string;
    descripcion: string;
    especial: string;
    etag: string;
    fechaCreacion: string;
    fechaModificacion: string;
    os: string;
    partitionKey: string;
    rowKey: string;
    size: string;
    timestamp: string;
    usuarioCreacion: string;
    usuarioModificacion: string;
}